+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/graph/IrGraphGenerator.kt
@@ -322,9 +322,10 @@ internal class IrGraphGenerator(
           graphMetadataReporter.write(node, bindingGraph)
           val metroMetadata = createMetroMetadata(dependency_graph = graphProto)
 
-          writeDiagnostic({
-            "graph-metadata-${node.sourceGraph.kotlinFqName.asString().replace(".", "-")}.kt"
-          }) {
+          writeDiagnostic(
+            "graph-metadata",
+            { "${node.sourceGraph.kotlinFqName.asString().replace(".", "-")}.kt" },
+          ) {
             metroMetadata.toString()
           }
 
@@ -749,10 +750,10 @@ internal class IrGraphGenerator(
       }
       .toList()
       .also { propertyBindings ->
-        writeDiagnostic("keys-providerProperties-${diagnosticTag}.txt") {
+        writeDiagnostic("keys-providerProperties", "${diagnosticTag}.txt") {
           propertyBindings.joinToString("\n") { it.binding.typeKey.toString() }
         }
-        writeDiagnostic("keys-scopedProviderProperties-${diagnosticTag}.txt") {
+        writeDiagnostic("keys-scopedProviderProperties", "${diagnosticTag}.txt") {
           propertyBindings
             .filter { it.binding.isScoped() }
             .joinToString("\n") { it.binding.typeKey.toString() }